Migrate flow attribute beans to use longs (#1062)
## Rationale
Flow attribute tables have integer `rowid` and `id` columns, but the
corresponding beans can work with longs.
https://github.com/LabKey/internal-issues/issues/1522

+++ flow/src/org/labkey/flow/query/KeywordForeignKey.java
@@ -69,7 +69,7 @@ public class KeywordForeignKey extends AttributeForeignKey<String>
     }
 
     @Override
-    protected SQLFragment sqlValue(ColumnInfo objectIdColumn, String attrName, int attrId)
+    protected SQLFragment sqlValue(ColumnInfo objectIdColumn, String attrName, long attrId)
     {
         // SQL server 2000 does not allow a TEXT column (i.e. flow.keyword.value) to appear in this subquery.
         // For this reason, we cast it to VARCHAR(4000).
